Refactoring and documentation improvements
