+++ metrics/web/test/auth/presentation/widgets/strategy/google_sign_in_option_appearance_strategy_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:metrics/auth/presentation/strings/auth_strings.dart';
-import 'package:metrics/auth/presentation/widgets/strategy/google_sign_in_option_strategy.dart';
+import 'package:metrics/auth/presentation/widgets/strategy/google_sign_in_option_appearance_strategy.dart';
 import 'package:metrics/common/presentation/button/theme/style/metrics_button_style.dart';
 import 'package:metrics/common/presentation/metrics_theme/model/login_theme_data.dart';
 import 'package:metrics/common/presentation/metrics_theme/model/metrics_theme_data.dart';
@@ -10,7 +10,7 @@ import 'package:test/test.dart';
 import '../../../../test_utils/auth_notifier_mock.dart';
 
 void main() {
-  group("GoogleSignInOptionStrategy", () {
+  group("GoogleSignInOptionAppearanceStrategy", () {
     const loginThemeData = LoginThemeData(
       loginOptionButtonStyle: MetricsButtonStyle(
         color: Colors.red,
@@ -50,19 +50,21 @@ void main() {
     });
 
     test(
-        ".getWidgetAppearance() returns the login option button theme if not in the loading state",
-        () {
-      final actualTheme = strategy.getWidgetAppearance(metricsTheme, false);
+      ".getWidgetAppearance() returns the login option button theme if not in the loading state",
+      () {
+        final actualTheme = strategy.getWidgetAppearance(metricsTheme, false);
 
-      expect(actualTheme, loginThemeData.loginOptionButtonStyle);
-    });
+        expect(actualTheme, loginThemeData.loginOptionButtonStyle);
+      },
+    );
 
     test(
-        ".getWidgetAppearance() returns the inactive login option button theme if in the loading state",
-        () {
-      final actualTheme = strategy.getWidgetAppearance(metricsTheme, true);
+      ".getWidgetAppearance() returns the inactive login option button theme if in the loading state",
+      () {
+        final actualTheme = strategy.getWidgetAppearance(metricsTheme, true);
 
-      expect(actualTheme, loginThemeData.inactiveLoginOptionButtonStyle);
-    });
+        expect(actualTheme, loginThemeData.inactiveLoginOptionButtonStyle);
+      },
+    );
   });
 }
